Move SearchResult to the Search screen folder
